Adding support for nested pages.
The routing allowed for nested pages (/some/path/here), but the controller would not only not look
up the page with the handle of 'some/path/here', but would actually look up a different page (just
the first part, or 'some' in this case). This meant you could create pages that couldn't be seen
later.