Bugfix: letting Snippets be edited.
Adding Snippet#full_contents.
Snippet will do no processing (full_contents is an alias for contents),
but it's troublesome to *not* have full_contents, especially when it comes
to previewing.

Maybe later snippets will include other snippets, but only when it's needed
and only if we have a good strategy for avoiding loops.