Fixing layout-dependent PagesController specs.
We spec under a Rails app.  If that app is under any development it presumably has a
layout, etc.  If it's a new app (like I just created) then it will not have an application
layout file.  Making these specs more robust so that they can detect if the "default" layout
is in use, whether there is an application layout present or not.