Adding replacement val inputs to Page admin forms
This only includes what the associate template has as replacements at
the time of form page loading. This means page creation could have an
extra step if a template is involved, or it could mean nesting page
creation under template is wanted.

We'll figure it out.

Also, the page.value ||= {} line is ugly. Page#values should probably
default to an empty hash.

Closes https://www.pivotaltracker.com/story/show/18379457