Adding format to layout admin form.
Some trickiness because I want to promote HAML over ERB,
but Rails uses ERB as default (as in something with no extension is
treated as ERB).

This solution actually makes ERB the default for new layouts as well,
since I didn't want to get even nastier with the form to base the
default on context.