Allowing Template to have multiple formats.
It's really HAML/ERB, but the ERB will be handled automatically by
Rails. And the defaults might be a little wrong. And formatting might
clash between snippets and pages and templates when combined in certain
ways.

And maybe it needs Markdown as a format as well. We'll jump off that
bridge when we get to it.

Maybe there should be a site-wide config area and setting for things
like this. Another bridge-jumping.