Making HAML-layout snippets nicer.
For HAML layouts, the indentation levels are very important. Snippets
were inserted as-is, which meant you had to know about the layout to
make the snippet. This indents the snippet to the replacement, so
everything is hunky-dory.

Didn't want to muck with expansion for other types, so inserted this
directly in the Layout class.